+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - creating of private/public statics
 - render list of public statics at statics page
 
+### Modified
+- allow to RL and CL of guild approve characters in guild statics
+
 ## [0.8.1] - 2019-11-17
 ### Fixed
 - bug with event creation for statics